+++ src/actions/category/IndexAction.php
@@ -15,6 +15,7 @@
 namespace blackcube\admin\actions\category;
 
 use blackcube\admin\actions\BaseElementAction;
+use blackcube\core\interfaces\PluginsHandlerInterface;
 use blackcube\core\models\Category;
 use blackcube\core\models\Slug;
 use blackcube\core\models\Type;
@@ -83,7 +84,11 @@ class IndexAction extends BaseElementAction
                 ]
             ],
         ]);
+        $pluginsHandler = Yii::createObject(PluginsHandlerInterface::class);
+        /* @var $pluginsHandler \blackcube\core\interfaces\PluginsHandlerInterface */
+
         return $this->controller->render($this->view, [
+            'pluginsHandler' => $pluginsHandler,
             'categoriesProvider' => $categoriesProvider,
         ]);
     }