+++ src/actions/composite/DeleteAction.php
@@ -16,6 +16,8 @@ namespace blackcube\admin\actions\composite;
 
 use blackcube\admin\actions\BaseElementAction;
 use blackcube\admin\Module;
+use blackcube\core\interfaces\PluginHookInterface;
+use blackcube\core\interfaces\PluginsHandlerInterface;
 use yii\web\NotFoundHttpException;
 use yii\web\Response;
 use Yii;
@@ -54,8 +56,16 @@ class DeleteAction extends BaseElementAction
         }
         if (Yii::$app->request->isPost) {
             $transaction = Module::getInstance()->db->beginTransaction();
+            $pluginsHandler = Yii::createObject(PluginsHandlerInterface::class);
+            /* @var $pluginsHandler \blackcube\core\interfaces\PluginsHandlerInterface */
+
             try {
                 $slug = $composite->getSlug()->one();
+                $deletePlugins = $pluginsHandler->runHook(PluginHookInterface::PLUGIN_HOOK_DELETE, $composite);
+                $deletePlugins = array_reduce($deletePlugins, function($accumulator, $item) {
+                    return $accumulator && $item;
+                }, true);
+
                 if ($slug !== null) {
                     $slug->delete();
                 }