+++ src/actions/node/IndexAction.php
@@ -15,6 +15,7 @@
 namespace blackcube\admin\actions\node;
 
 use blackcube\admin\actions\BaseElementAction;
+use blackcube\core\interfaces\PluginsHandlerInterface;
 use blackcube\core\models\Node;
 use blackcube\core\models\Slug;
 use blackcube\core\models\Type;
@@ -86,8 +87,11 @@ class IndexAction extends BaseElementAction
                 ]
             ],
         ]);
+        $pluginsHandler = Yii::createObject(PluginsHandlerInterface::class);
+        /* @var $pluginsHandler \blackcube\core\interfaces\PluginsHandlerInterface */
 
         return $this->controller->render($this->view, [
+            'pluginsHandler' => $pluginsHandler,
             'nodesProvider' => $nodesProvider
         ]);
     }