+++ src/actions/plugin/IndexAction.php
@@ -1,0 +1,59 @@
+<?php
+/**
+ * IndexAction.php
+ *
+ * PHP version 7.2+
+ *
+ * @author Philippe Gaultier <pgaultier@redcat.io>
+ * @copyright 2010-2020 Redcat
+ * @license https://www.redcat.io/license license
+ * @version XXX
+ * @link https://www.redcat.io
+ * @package blackcube\admin\actions\plugin
+ */
+
+namespace blackcube\admin\actions\plugin;
+
+use blackcube\core\components\Plugins;
+use blackcube\core\helpers\PluginHelper;
+use blackcube\core\interfaces\PluginsHandlerInterface;
+use blackcube\core\models\Plugin;
+use blackcube\core\Module as CoreModule;
+use yii\base\Action;
+use yii\base\InvalidArgumentException;
+use yii\web\NotFoundHttpException;
+use yii\web\Response;
+use Yii;
+
+/**
+ * Class IndexAction
+ *
+ * @author Philippe Gaultier <pgaultier@redcat.io>
+ * @copyright 2010-2020 Redcat
+ * @license https://www.redcat.io/license license
+ * @version XXX
+ * @link https://www.redcat.io
+ * @package blackcube\admin\actions\type
+ */
+class IndexAction extends Action
+{
+    /**
+     * @var string view
+     */
+    public $view = 'index';
+
+    /**
+     * @return string|Response
+     * @throws NotFoundHttpException
+     * @throws \yii\base\InvalidConfigException
+     */
+    public function run()
+    {
+        $pluginsHandler = Yii::createObject(PluginsHandlerInterface::class);
+        /* @var $pluginsHandler \blackcube\core\interfaces\PluginsHandlerInterface */
+        $pluginManagers = $pluginsHandler->getPluginManagers();
+        return $this->controller->render($this->view, [
+            'pluginManagers' => $pluginManagers,
+        ]);
+    }
+}