+++ src/actions/tag/DeleteAction.php
@@ -18,6 +18,8 @@ use blackcube\admin\actions\BaseElementAction;
 use blackcube\admin\Module;
 use blackcube\core\models\Tag;
 use yii\base\Action;
+use blackcube\core\interfaces\PluginHookInterface;
+use blackcube\core\interfaces\PluginsHandlerInterface;
 use yii\web\NotFoundHttpException;
 use yii\web\Response;
 use Yii;
@@ -55,8 +57,16 @@ class DeleteAction extends BaseElementAction
         }
         if (Yii::$app->request->isPost) {
             $transaction = Module::getInstance()->db->beginTransaction();
+            $pluginsHandler = Yii::createObject(PluginsHandlerInterface::class);
+            /* @var $pluginsHandler \blackcube\core\interfaces\PluginsHandlerInterface */
+
             try {
                 $slug = $tag->getSlug()->one();
+                $deletePlugins = $pluginsHandler->runHook(PluginHookInterface::PLUGIN_HOOK_DELETE, $tag);
+                $deletePlugins = array_reduce($deletePlugins, function($accumulator, $item) {
+                    return $accumulator && $item;
+                }, true);
+
                 if ($slug !== null) {
                     $slug->delete();
                 }