+++ src/actions/tag/IndexAction.php
@@ -15,6 +15,7 @@
 namespace blackcube\admin\actions\tag;
 
 use blackcube\admin\actions\BaseElementAction;
+use blackcube\core\interfaces\PluginsHandlerInterface;
 use blackcube\core\models\Category;
 use blackcube\core\models\Slug;
 use blackcube\core\models\Tag;
@@ -85,7 +86,10 @@ class IndexAction extends BaseElementAction
                 ]
             ],
         ]);
+        $pluginsHandler = Yii::createObject(PluginsHandlerInterface::class);
+        /* @var $pluginsHandler \blackcube\core\interfaces\PluginsHandlerInterface */
         return $this->controller->render('index', [
+            'pluginsHandler' => $pluginsHandler,
             'tagsProvider' => $tagsProvider
         ]);
     }