+++ src/views/node/index.php
@@ -11,6 +11,7 @@
  * @link https://www.redcat.io
  * @package blackcube\admin\views\node
  *
+ * @var $pluginsHandler \blackcube\core\interfaces\PluginsHandlerInterface
  * @var $nodesProvider \yii\data\ActiveDataProvider
  * @var $this \yii\web\View
  */
@@ -18,15 +19,38 @@
 use blackcube\admin\Module;
 use blackcube\admin\helpers\Html;
 use blackcube\admin\widgets\Sidebar;
+use blackcube\core\interfaces\PluginsHandlerInterface;
+use blackcube\admin\interfaces\PluginAdminHookInterface;
 use yii\widgets\LinkPager;
 
 $formatter = Yii::$app->formatter;
 ?>
     <main class="overflow-hidden">
+        <?php if ($pluginsHandler instanceof PluginsHandlerInterface): ?>
+            <?php $widgets = $pluginsHandler->runWidgetHook(PluginAdminHookInterface::PLUGIN_HOOK_WIDGET_NODE_LIST_HEAD); ?>
+            <?php foreach ($widgets as $widget): ?>
+                <?php echo $widget; ?>
+            <?php endforeach; ?>
+        <?php endif; ?>
         <div class="table-container">
+            <?php if ($pluginsHandler instanceof PluginsHandlerInterface): ?>
+                <?php $widgets = $pluginsHandler->runWidgetHook(PluginAdminHookInterface::PLUGIN_HOOK_WIDGET_NODE_LIST_BEFORE_LIST); ?>
+                <?php foreach ($widgets as $widget): ?>
+                    <?php echo $widget; ?>
+                <?php endforeach; ?>
+            <?php endif; ?>
             <div blackcube-ajaxify="click" blackcube-attach-modal="">
-                <?php echo $this->render('_list', ['nodesProvider' => $nodesProvider]); ?>
+                <?php echo $this->render('_list', [
+                    'pluginsHandler' => $pluginsHandler,
+                    'nodesProvider' => $nodesProvider
+                ]); ?>
             </div>
+            <?php if ($pluginsHandler instanceof PluginsHandlerInterface): ?>
+                <?php $widgets = $pluginsHandler->runWidgetHook(PluginAdminHookInterface::PLUGIN_HOOK_WIDGET_NODE_LIST_AFTER_LIST); ?>
+                <?php foreach ($widgets as $widget): ?>
+                    <?php echo $widget; ?>
+                <?php endforeach; ?>
+            <?php endif; ?>
             <div class="text-center">
                 <?php echo LinkPager::widget([
                     'pagination' => $nodesProvider->pagination,