+++ src/views/tag/form.php
@@ -11,6 +11,7 @@
  * @link https://www.redcat.io
  * @package blackcube\admin\views\tag
  *
+ * @var $pluginsHandler \blackcube\core\interfaces\PluginsHandlerInterface
  * @var $tag \blackcube\core\models\Tag
  * @var $slugForm \blackcube\admin\models\SlugForm
  * @var $categoriesQuery \blackcube\core\models\FilterActiveQuery
@@ -24,12 +25,23 @@ use blackcube\admin\Module;
 use blackcube\admin\helpers\Html;
 use blackcube\admin\widgets\Sidebar;
 use blackcube\admin\widgets\SlugForm;
+use blackcube\core\interfaces\PluginsHandlerInterface;
+use blackcube\admin\interfaces\PluginAdminHookInterface;
 use yii\helpers\ArrayHelper;
 use yii\helpers\Url;
 
 ?>
     <main>
         <?php echo Html::beginForm('', 'post', ['class' => 'form']); ?>
+        <?php if ($pluginsHandler instanceof PluginsHandlerInterface): ?>
+            <?php $widgets = $pluginsHandler->runWidgetHook(
+                $tag->isNewRecord ? PluginAdminHookInterface::PLUGIN_HOOK_WIDGET_TAG_CREATE_HEAD : PluginAdminHookInterface::PLUGIN_HOOK_WIDGET_TAG_UPDATE_HEAD,
+                $tag
+            ); ?>
+            <?php foreach ($widgets as $widget): ?>
+                <?php echo $widget; ?>
+            <?php endforeach; ?>
+        <?php endif; ?>
         <ul class="header">
             <li>
                 <?php echo Html::a('<i class="fa fa-angle-left mr-2"></i> '.Module::t('tag', 'Back'), ['index'], ['class' => 'button']); ?>
@@ -85,6 +97,15 @@ use yii\helpers\Url;
                     <?php echo Html::activeTextInput($tag, 'name', ['class' => 'textfield'.($tag->hasErrors('name')?' error':'')]); ?>
                 </div>
             </div>
+            <?php if ($pluginsHandler instanceof PluginsHandlerInterface): ?>
+                <?php $widgets = $pluginsHandler->runWidgetHook(
+                    $tag->isNewRecord ? PluginAdminHookInterface::PLUGIN_HOOK_WIDGET_TAG_CREATE_BEFORE_BLOCS : PluginAdminHookInterface::PLUGIN_HOOK_WIDGET_TAG_UPDATE_BEFORE_BLOCS,
+                    $tag
+                ); ?>
+                <?php foreach ($widgets as $widget): ?>
+                    <?php echo $widget; ?>
+                <?php endforeach; ?>
+            <?php endif; ?>
 
             <?php if ($tag->id !== null && $tag->type !== null): ?>
                 <?php echo Html::beginTag('div', ['blackcube-toggle-element' => Html::bindAureliaAttributes([
@@ -123,6 +144,16 @@ use yii\helpers\Url;
                         </div>
                     <?php echo Html::endTag('div'); ?>
                 <?php echo Html::endTag('div'); ?>
+            <?php endif; ?>
+
+            <?php if ($pluginsHandler instanceof PluginsHandlerInterface): ?>
+                <?php $widgets = $pluginsHandler->runWidgetHook(
+                    $tag->isNewRecord ? PluginAdminHookInterface::PLUGIN_HOOK_WIDGET_TAG_CREATE_TAIL : PluginAdminHookInterface::PLUGIN_HOOK_WIDGET_TAG_UPDATE_TAIL,
+                    $tag
+                ); ?>
+                <?php foreach ($widgets as $widget): ?>
+                    <?php echo $widget; ?>
+                <?php endforeach; ?>
             <?php endif; ?>
 
             <div class="buttons">